Move some types from AbstractContentContext.h to ContentBasicTypes.h
