+++ PDFWriter/AbstractContentContext.h
@@ -37,8 +37,7 @@
 #include "GraphicStateStack.h"
 #include "GlyphUnicodeMapping.h"
 #include "ObjectsBasicTypes.h"
-#include <string>
-#include <list>
+#include "ContentBasicTypes.h"
 #include <set>
 #include <utility>
 
@@ -46,7 +45,8 @@
 namespace PDFHummus
 {
     class DocumentContext;  
-};
+}
+
 class ObjectsContext;
 class PDFStream;
 class ResourcesDictionary;
@@ -55,22 +55,8 @@ class ITextCommand;
 class IByteReader;
 class IContentContextListener;
 
-template <typename T>
-struct SomethingOrDouble
-{
-	T SomeValue;
-	double DoubleValue;
-
-	bool IsDouble; // true - double, false - other
-
-	SomethingOrDouble(T inSome){SomeValue = inSome;IsDouble = false;}
-	SomethingOrDouble(double inDouble){DoubleValue = inDouble;IsDouble = true;}
-};
-
-typedef SomethingOrDouble<std::string> StringOrDouble;
 typedef SomethingOrDouble<GlyphUnicodeMappingList> GlyphUnicodeMappingListOrDouble;
 
-typedef std::list<StringOrDouble> StringOrDoubleList;
 typedef std::list<GlyphUnicodeMappingListOrDouble> GlyphUnicodeMappingListOrDoubleList;
 
 typedef std::set<IContentContextListener*> IContentContextListenerSet;
