+++ PDFWriter/ContentBasicTypes.h
@@ -1,0 +1,40 @@
+/*
+   Source File : ObjectsBasicTypes.h
+
+
+   Copyright 2011 Gal Kahana PDFWriter
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+
+   
+*/
+#pragma once
+
+#include <string>
+#include <list>
+
+template <typename T>
+struct SomethingOrDouble
+{
+	T SomeValue;
+	double DoubleValue;
+
+	bool IsDouble; // true - double, false - other
+
+	SomethingOrDouble(T inSome){SomeValue = inSome;IsDouble = false;}
+	SomethingOrDouble(double inDouble){DoubleValue = inDouble;IsDouble = true;}
+};
+
+typedef SomethingOrDouble<std::string> StringOrDouble;
+
+typedef std::list<StringOrDouble> StringOrDoubleList;
